Add Money.__round__ dunder method
Adds support for calling ``round()`` on Money instances.

The implementation uses Decimal.quantize under the hood to
ensure stable selection of rounding mode, and allows
the user to select their preferred mode by wrapping the call
in a decimal context suiting their needs.

Calling round with no arguments raises an ``TypeError``
as it would drop the currency information contained in the money class.